[NAE-1947] HistoryService findAllSetDataEventLogs is not working - change access modifier in ChangedField - move ChangedField to another package - remove unnecessary comments
