+++ src/main/groovy/com/netgrif/application/engine/petrinet/domain/dataset/logic/ChangedFieldsTree.groovy
@@ -1,6 +1,6 @@
 package com.netgrif.application.engine.petrinet.domain.dataset.logic
 
-import com.netgrif.application.engine.petrinet.domain.ChangedField
+
 import com.netgrif.application.engine.workflow.domain.Task
 import com.querydsl.core.annotations.QueryExclude
 import lombok.Getter
