+++ src/main/java/com/netgrif/application/engine/history/domain/dataevents/SetDataEventLog.java
@@ -2,7 +2,7 @@ package com.netgrif.application.engine.history.domain.dataevents;
 
 import com.netgrif.application.engine.auth.domain.IUser;
 import com.netgrif.application.engine.history.domain.taskevents.TaskEventLog;
-import com.netgrif.application.engine.petrinet.domain.ChangedField;
+import com.netgrif.application.engine.petrinet.domain.dataset.logic.ChangedField;
 import com.netgrif.application.engine.petrinet.domain.events.EventPhase;
 import com.netgrif.application.engine.workflow.domain.Case;
 import com.netgrif.application.engine.workflow.domain.Task;
