+++ src/main/java/com/netgrif/application/engine/workflow/domain/eventoutcomes/dataoutcomes/SetDataEventOutcome.java
@@ -1,6 +1,6 @@
 package com.netgrif.application.engine.workflow.domain.eventoutcomes.dataoutcomes;
 
-import com.netgrif.application.engine.petrinet.domain.ChangedField;
+import com.netgrif.application.engine.petrinet.domain.dataset.logic.ChangedField;
 import com.netgrif.application.engine.workflow.domain.Case;
 import com.netgrif.application.engine.workflow.domain.Task;
 import com.netgrif.application.engine.workflow.domain.eventoutcomes.EventOutcome;
